refactor: Change file names to index.js
